Move to one consumer
