+++ parflow_data_management/scheduler/templates/execute.html
@@ -15,7 +15,7 @@
   const commandSocket = new WebSocket(
       'ws://'
       + window.location.host
-      + '/ws/execute/'
+      + '/ws/data/'
   );
 
   document.getElementById("execute_btn").onclick = function() {
