+++ parflow_data_management/transport/templates/basicbutton.html
@@ -15,7 +15,7 @@
   const keySocket = new WebSocket(
       'ws://'
       + window.location.host
-      + '/ws/keygen/'
+      + '/ws/data/'
   );
 
   document.getElementById("generate_button_id").onclick = function() {
